refactor(ts/components/map): move street view control into own file
